requirements.txt
django==1.11.15
djangorestframework==3.9.0
markdown==3.0.1
psycopg2-binary==2.7.6.1
mongoengine==0.9.0
pymongo==2.8
